decomposition room into components
