+++ project/src/components/room-host/room-host.tsx
@@ -1,0 +1,28 @@
+import { Host } from '../../types/offer';
+
+type RoomHostProps = {
+  host: Host;
+  description: string;
+}
+
+function RoomHost({host, description}: RoomHostProps): JSX.Element {
+  const {name, avatarUrl, isPro} = host;
+
+  return (
+    <div className="property__host">
+      <h2 className="property__host-title">Meet the host</h2>
+      <div className="property__host-user user">
+        <div className="property__avatar-wrapper property__avatar-wrapper--pro user__avatar-wrapper">
+          <img className="property__avatar user__avatar" src={avatarUrl} width="74" height="74" alt={name} />
+        </div>
+        <span className="property__user-name">{name}</span>
+        {isPro && <span className="property__user-status">Pro</span>}
+      </div>
+      <div className="property__description">
+        <p className="property__text">{description}</p>
+      </div>
+    </div>
+  );
+}
+
+export default RoomHost;
